requirements.txt
google_api_python_client==2.146.0
oauth2client==4.1.3
Pillow==11.1.0
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
python_dateutil==2.9.0.post0
stressrnn==0.2.2
